Rebuilt add mode
For a better user experience, the add mode has been rewritten so that it doesn't require the user to continuously write /add in front of stuff.

From now on, the user enters the add mode with /add, then send stuff they want to add as messages, and exit the add mode with /done.